Successfully moved CSS into external files.
Currently all of the stylesheets will be rendered into the "Blue.tmpl" file,
which will provide the structure.  In the future the extra .tmpl files could
be removed, or the user could be offered the selection of different types
of HTML structure divorced from the different types of CSS structure.