Provided a CMake variable BIBLETIME_FRONTEND to select the frontend to build.
To add a new frontend, the bibletime_FRONTEND_${BIBLETIME_FRONTEND}_SOURCES
and bibletime_FRONTEND_${BIBLETIME_FRONTEND}_MOCABLE_HEADERS variables must be
set in cmake/BTSourceFileList.cmake, and a new frontend string must be added
to bibletime_AVAILABLE_FRONTENDS in the same file.

Note that currently this does not affect translations, which are still
generated from all the sources. Support for translations must be added later.
This can and must be done without adding new translation files for each
frontend.