Fixed extra *.tmpl files not being considered as template candidates.
Now all *.css and *.tmpl files in the display-templates/ folder are
considered as display templates with the exception of "Basic.tmpl",
which is currently the base for all *.css templates. This finally
provides the ability for templates to provide extra functionality
via javascript inclusion etc.

This was a port from stable-2.9.

Conflicts:
	src/backend/managers/cdisplaytemplatemgr.cpp