backend: Optimized deleting orphaned of indexes and probably fixed a crash.
This code may crash because we were using operator[], directly on QDir, for
which to the documentation does not guarantee an entry will exist after the
previous call to count(), possible leading to and invalid index assertion.
Instead we now first create a constant copy of the entry list, and iterate on
the copy instead.