Refactored BtInstallThread and the install progress dialog.
This was in part done to fix a bug which caused BibleTime to freeze if the
installation of a module failed. This was caused by an unhandled case in
BtInstallThread which resulted in BtInstallProgressDialog not being notified
of failure.

While working out this bug, numerous other bugs were found and I refactored the
classes a bit. Now a single BtInstallThread thread handles installation of all
modules. I also removed buttons to stop installation of single modules from the
progress dialog. This made the design much more simple, and since installation
was not done in parallel anyway, these buttons had ambigious semantics.