frontend, BibleTime: Work around moc failure in Release mode.
The moc_bibletime.cpp file generated in release mode still referred to the
slotDebugWindowClosing, slotDebugTimeout and slotShowDebugWindow methods.
This caused compilation of moc_bibletime.cpp to fail because these methods
are not declared when NDEBUG is set.

The actual reason for this is that moc is not given the same -DNDEBUG define.
I'm currently too lazy to fix this properly in our CMake files, since using
C++11-style connect/disconnect is something we need to switch to in the future
anyway.

Thanks to Martin Gruner for reporting this issue.