Hallelujah - Initial working zoom in/out/reset menu options in Lexicon Display
✝ I wanted to be able to zoom out on large images and maps. This is my first time working
on a Qt project communally thanks. I code with _aleluya added to variables/etc...
It reminds me and hopefully others to praise Jesus. He gives the talent to code.

The zoomer right now is simple, only doing a *2 or /2 and reset to 1 by using the
zoomFactor of QWebView, the method of accessing this functionality is understood
to be ok, as it was used from the raw html functions available.

I am not sure of any consequences yet of the HTML web engine functionality.
Plans include: add to Bible view, spanish and other translations, add finer
grained zoom, perhaps toolbar buttons and zoom around where mouse is centered,
use CTRL-mousewheel to zoom. I am not sure i will finish this as the main
functionality i want is there but God willing i will.

God guide us and bless you in Jesus Holy name