.travis.yml: Removed all arm64 jobs
Supporting arm64 jobs can not be done in a straightforward manner, i.e. by using
the APT addon. We could probably use before_script or script to add repositories
and install packages based on the architecture or environment, but this would
make the .travis.yml file quite a hack. Hence it seems better just to omit the
arm64 builds altogether.