Use #pragma once instead of header guards
All modern C++ compilers seem to support #pragma once, and this results in much
cleaner code and slightly faster compilation. The macros we used as header
guards were also inconsistent in their naming, so we get rid of that as well.

This commit only leaves platforms/common/clucene/CLucene/clucene-config.h with
header guards, as the macro those define might have a special meaning for
Clucene.