frontend, CDisplayWindow: Fixed incorrect cast in storeProfileSettings()
This cast silently become invalid when sword::VerseKey ceased to be used as a
base class of CSwordVerseKey (commit 4c77038ef35ada969715d4946b94704e9c47efc7).
This was an undetected use of the previous design pattern.

The fix is to directly attempt a cast to CSwordVerseKey, on which then to
operate. Also, we no longer mutate the original *m_swordKey, object but instead
clone from it a temporary copy for modification (using the setLocale() method
introduced by the previous commit).