frontend, CDisplaySettings: Optimized iterating over available locales
Since sword::LocaleMgr::getAvailableLocales() always returns a list of locale
abbreviations, it does not make much sense to convert these back to SWLocale
references with sword::LocaleMgr::getLocale() before re-reading the abbreviation
from SWLocale::getName().

It is unfortunate that Sword does not directly provide a list of locales instead
of the list of locale abbreviations, as providing the latter in not optimal.