backend, CLanguageMgr: Optimized Language abbrevation handling and lookup
Since LanguageInfo fields m_langMap and m_langList were only used for lookup of
languages based on their abbreviations in a single lookup phase, and since there
is no overlap between the primary and alternative abbreviations, it's reasonable
to just use a single hashmap for them.

This also simplifies Language construction, whereby the whole list of
abbreviations applicable for the language are given as one argument to the
constructor, and is also handled as a single class field.