Add text-to-speech (TTS) functionality for currently selected text
This adds a basic text-to-speech (TTS) functionality for
the currently selected text, using QTextToSpeech.

Speech output can be triggered as follows:

* select text
* open context menu buy right-clicking on the selected text
* select the newly added  "Speak selected text" entry

In order to have QTextToSpeech use the proper language,
a locale generated from the text language is set.

Currently, only the locale is explicitly set for QTextToSpeech,
which means that the default voice for the given locale is used.
(This could be made configurable in a follow-up step to e.g.
allow the user to specify what voice to use for what language in
the application settings.)

The newly needed Qt5TextToSpeech module is added in
cmake/BTApplication.cmake accordingly.

Related: Issue #5