backend, ThMLtoHTML: Fixed assertion error in optional <sync> attribute handling
This is a fixup to ba09655dfe32e285614f856806279ff5d3f9f60f, in which I
erronously added the incorrect BT_ASSERT() call. I failed to note that the
"class" attribute (which is actually a Sword addition rather than being
specified in ThML) is optional and that the previous logic actually relied on
QRegExp::cap() to return empty strings for captures not matching anything.