frontend, module choosers: Use pointers for handling modules
BtWindowModuleChooser, its subclasses and the rest of the BtModuleChooser* and
BtTextWindow* classes used to manage modules by their names, using QStringList
and QString. This is both inefficient and cumbersome, which often requires the
actual CSwordModuleInfo classes to be looked up from the backend by name. It is
better to handle modules using BtModuleList and raw CSwordModuleInfo * pointers
instead.