frontend: Reimplemented much of the text selection code, fixing some bugs
The previous implementation contained a number of bugs, including selections
spanning multiple rows to be incorrect etc. Since the code turned out to be very
difficult for me to understand and fix more cleanly I decided to rewrite the
relevant parts.

This is a fairly large commit, as the selection code spans multiple components,
but mostly QML and QML-related code. Due to changes to the way BtQmlInterface
keeps track of selections, the getters returning indexes of selected items were
replaced with a more logical selection() getter. This incurred some small but
healthy changes to how BtCopyByReferencesDialog uses the selection indexes.