backend, CSwordModuleSearch: Refactored to remove confusing check
Looking at source code of both BibleTime and SWORD (as of SVN 3887) it seems
that the only way to crash due to this is to incorrectly static_cast SWKey
instances to sword::VerseKey. By instead switching to a dynamic_cast in
CSwordModuleInfo::searchIndexed() we can get rid of this check. Let's hope it
works.