backend, CSwordKey: Fixed module key setting if different locale
If the locale of the module key and this key did not match, the module key did
not properly get set using text only, leading to bizarre bugs.

The specific bug which led me to this discovery was related to editing the
writable "personal" module: when editing the same verse two times in a row then
the second time BtEditTextWizard showed up with an empty text, which was I think
was because the module key did not properly change, but pointed to an empty
verse entry.