backend, CSwordLexiconModuleInfo: Fixed cache QDataStream compatibility issue
The QDataStream version was not fixed before serialization and deserialization of
the first values on the stream, potentially resulting in errors during cache
deserialization which were not immediately detected, leading to issues elsewhere.

This commit also bumps the cache format, introduces updating the cache file
whenever an older format is read, and improves debug output.