frontend, BibleTimeApp: Fix and refactor sword.conf initialization on Win/Mac
This was broken by commit c9b99076a501, which used copy-paste to inline the code
and resulted the return statement to become incorrect code, as it exited the
function it was inlined into, not just skipping the rest of the inlined code.

The same commit also incorrectly extended the scope of the QFile object,
causing it not being closed correctly early on.

This commit fixes these issues, and simplifies the code.