Don't offer copying selection when there is none
The popup menu shown when right-clicking on a Bible text
has a "Copy" -> "Copy" entry that copies the currently selected
text.

When there is no selection, that entry would previously copy
an empty string to the clipboard.

Instead, adjust the logic so that the menu entry is disabled if there
is no text selection.

Set the action only initially to enabled if there is a selection,
otherwise disable it.

So far, the BtQmlInterface::selectionChanged signal was
only emitted in BtQmlInterface::setSelection when a new (non-empty)
selection gets set, but not in BtQmlInterface::clearSelection when
the selection gets cleared, meaning that the existing slot connected
to that signal would never disable the action.

Emit the signal in BtQmlInterface::clearSelection with an empty
std::optional for the new selection so that the slot now also
disables the menu entry again when the selection gets cleared.

Sample scenario with this commit in place:

1) Start BibleTime
2) open the popup menu (without any selection)
-> "Copy" -> "Copy" entry is disabled
3) select text
4) open popup menu
-> "Copy" -> "Copy" entry is enabled, i.e. can be selected
5) select "Copy" -> "Copy" menu entry
-> currently selected text gets copied to the clipboard
6) deselect text again
7) open the popup menu (without any selection)
-> "Copy" -> "Copy" entry is disabled

At a quick glance, it looks like the behavior should
have been the same before

    commit 87fda68fa569a73e86ffa95bd2a68c5f37bb7b26
    Date:   Sat Jun 29 23:59:19 2024 +0300

        frontend, C(Display|BibleRead)Window: Avoid use of QMenu::aboutToShow()

, but the commit before that one fails to build locally for me,
so I didn't double-check.