Add text-to-speech (TTS) functionality for currently selected text
This adds basic text-to-speech (TTS) functionality for
the currently selected text, using QTextToSpeech.

Speech output can be triggered as follows:

* select text
* open context menu by right-clicking on the selected text
  or using the context menu key
* select the newly added "Speak selected text" context menu
  entry

The TTS engine, locale/language and voice can be configured
in a newly added "Text-to-speech" page in the "Settings" ->
"Configure BibleTime..." dialog.

Engines are provided by corresponding QTextToSpeech modules.
On Debian, two of them can e.g. be installed via the packages
`qt6-speech-flite-plugin` and  `qt6-speech-speechd-plugin`.
If no text-to-speech module is found, the text-to-speech settings
page will show a corresponding note.

The newly added icon `pics/icons/speaker.svg` used for that settings
page is a public domain icon retrieved from [1], resaved with
Inkscape-specific metdata removed. (The original file was using a
format that is called "Inkscape SVG" in Inkscape's UI; use "Plain SVG"
instead to e.g. drop the setting what file name and path to use when
exporting to PNG, which was using a hard-coded path underneath
/home/john/).

The text-to-speech feature requires Qt's TextToSpeech module.
The feature is enabled by default, but can be disabled at build
time using the CMake switch `-DBUILD_TEXT_TO_SPEECH=OFF`.

Fixes: #5

[1] https://commons.wikimedia.org/wiki/File:Wikisource_speaker_icon.svg